se eliminó el archivo poligono de validaciones y se creo el archivo ValidaPoligonos para sus respec tivas validaciones
