+++ backend/package.json
@@ -18,7 +18,7 @@
     "luxon": "^3.4.0",
     "node-fetch": "^3.3.2",
     "nodemon": "^3.0.1",
-    "pg": "^8.11.1",
+    "pg": "^8.11.3",
     "validator": "^13.11.0"
   }
 }
